+++ constants/__init__.py
@@ -9,6 +9,10 @@ MAX_NOTIFICATION_LENGTH = 128
 
 ONLINE_RECENTLY_SECONDS = 12 * 60 * 60  # 12 hours
 
+# Most devices a person may stay signed in on at once; older sessions are
+# signed out on each new sign-in.
+MAX_SIGNED_IN_SESSIONS = 100
+
 # Club SEO page tunables. Shared by service/person/sql (API reads) and
 # service/cron/clubseo/sql (cron aggregation); kept in this dependency-free
 # module so both sides can import them without pulling each other in.
